gemini_chatbot/requirements.txt
google-generativeai>=0.5.2
python-dotenv>=1.0.0
google-search-results>=2.4.2
